LayoutTests/inspector/unit-tests/throttle-expected.txt

Test throttle proxy support.


== Running test suite: Throttle
-- Running test case: Basic throttle
Call throttled function every 1 ms for 110 ms.
PASS: All calls delayed at least 50 ms.

-- Running test case: Throttle proxy uniqueness
PASS: Each call to throttle should return a new proxy.

-- Running test case: Throttled function arguments
PASS: Trailing call invoked with most recent arguments.

-- Running test case: Cancel throttle
Canceled throttled function.
PASS: Throttled function canceled.
